+++ scripts/cet/Data/WorldObjectItem.lua
@@ -1,0 +1,4 @@
+---@class WorldObjectItem
+---@field label string?
+---@field item RHT_InspectorTarget?
+local WorldObjectItem = {}